Lab01.1/ddm501-lab1-starter/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
scikit-surprise==1.1.3

# Validation
pydantic==2.5.2

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
